0.4.2 module menu, renaming of the components - sub directories: buttons, forms, panels - module menu renamed to menuitem - UserController: new login(), logout(),   editCurrentUser() - backend.blade.php: logout is now a link   (not a button) - XxxController::routes():   usage of middleware('auth') - Auth::routes() removed

+++ templates/Http/Controllers/UserController.php
@@ -2,7 +2,6 @@
 
 namespace App\Http\Controllers;
 
-use App\Helpers\ContextLaraKnife;
 use App\Models\User;
 use App\Helpers\DbHelper;
 use App\Models\SProperty;
@@ -10,10 +9,13 @@ use App\Helpers\Pagination;
 use App\Helpers\ViewHelper;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
+use App\Helpers\ContextLaraKnife;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
 use Illuminate\Support\Facades\Redirect;
 use Illuminate\Support\Facades\Validator;
+use Symfony\Component\HttpFoundation\RedirectResponse;
 
 class UserController extends Controller
 {
@@ -52,12 +54,22 @@ class UserController extends Controller
         return $rc;
     }
     /**
+     * Shows the form for changing the password for the current user.
+     */
+    public function editCurrentUser(Request $request){
+        $user = User::find(auth()->id());
+        $rc = $this->editPassword($user, $request, '/menuitem-menu_main');
+        return $rc;
+    }
+
+
+    /**
      * Show the form for editing the specified resource.
      */
-    public function editPassword(User $user, Request $request)
+    public function editPassword(User $user, Request $request, ?string $redirect=null)
     {
         if ($request->btnSubmit === 'btnCancel') {
-            $rc = redirect('/user-index');
+            $rc = redirect($redirect ?? '/menuitem-menu_main');
         } else {
             $rc = null;
             $error = null;
@@ -70,7 +82,7 @@ class UserController extends Controller
                         'password_confirmation' => 'required'
                     ]);
                     $user->update($incomingFields);
-                    $rc = redirect('/user-index');
+                    $rc = redirect($redirect ?? '/user-index');
                 } catch (\Exception $e) {
                     $error = $e->getMessage();
                 }
@@ -126,6 +138,37 @@ class UserController extends Controller
         }
         return $rc;
     }
+    public function login(Request $request)
+    {
+        $rc = null;
+        $fields = $request->all();
+        if (count($fields) === 0) {
+            $fields = ['email' => '', 'password' => ''];
+        } else {
+            $credentials = $request->validate([
+                'email' => ['required', 'email'],
+                'password' => ['required'],
+            ]);
+            if (Auth::attempt($credentials)) {
+                $request->session()->regenerate();
+                $rc = redirect('/menuitem-menu_main');
+            } else {
+                $rc = back()->withErrors([
+                    'email' => __('The provided credentials do not match our records.'),
+                ])->onlyInput('email');
+            }
+        }
+        if ($rc == null) {
+            $context = new ContextLaraKnife($request, $fields);
+            $rc = view('user.login', ['context' => $context]);
+        }
+        return $rc;
+    }
+    public function logout(Request $request)
+    {
+        Auth::logout();
+        return redirect('/user-login');
+    }
     /**
      * Returns the validation rules.
      * @return array<string, string> The validation rules.
@@ -151,18 +194,26 @@ class UserController extends Controller
     }
     public static function routes()
     {
-        Route::get('/user-index', [UserController::class, 'index']);
-        Route::post('/user-index', [UserController::class, 'index']);
-        Route::get('/user-create', [UserController::class, 'create']);
-        Route::post('/user-create', [UserController::class, 'create']);
-        Route::put('/user-store', [UserController::class, 'store']);
-        Route::get('/user-edit/{user}', [UserController::class, 'edit']);
-        Route::post('/user-edit/{user}', [UserController::class, 'edit']);
-        Route::post('/user-update/{user}', [UserController::class, 'update']);
+        Route::get('/user-index', [UserController::class, 'index'])->middleware('auth');
+        Route::post('/user-index', [UserController::class, 'index'])->middleware('auth');
+        Route::get('/user-create', [UserController::class, 'create'])->middleware('auth');
+        Route::post('/user-create', [UserController::class, 'create'])->middleware('auth');
+        Route::put('/user-store', [UserController::class, 'store'])->middleware('auth');
+        Route::get('/user-edit/{user}', [UserController::class, 'edit'])->middleware('auth');
+        Route::post('/user-edit/{user}', [UserController::class, 'edit'])->middleware('auth');
+        Route::get('/user-edit-current', [UserController::class, 'editCurrentUser'])->middleware('auth');
+        Route::post('/user-edit-current', [UserController::class, 'editCurrentUser'])->middleware('auth');
+        Route::post('/user-update/{user}', [UserController::class, 'update'])->middleware('auth');
         Route::get('/user-editpassword/{user}', [UserController::class, 'editPassword']);
         Route::post('/user-editpassword/{user}', [UserController::class, 'editPassword']);
-        Route::get('/user-show/{user}/delete', [UserController::class, 'show']);
-        Route::delete('/user-show/{user}/delete', [UserController::class, 'destroy']);
+        Route::get('/user-show/{user}/delete', [UserController::class, 'show'])->middleware('auth');
+        Route::delete('/user-show/{user}/delete', [UserController::class, 'destroy'])->middleware('auth');
+        Route::get('/user-login', [UserController::class, 'login']);
+        Route::post('/user-login', [UserController::class, 'login']);
+        Route::get('/login', [UserController::class, 'login'])->name('login');;
+        Route::post('/login', [UserController::class, 'login'])->name('login');;
+        Route::get('/user-logout', [UserController::class, 'logout']);
+        Route::post('/user-logout', [UserController::class, 'logout']);
     }
     /**
      * Display the specified resource.
@@ -207,7 +258,9 @@ class UserController extends Controller
     public function update(User $user, Request $request)
     {
         $rc = null;
-        if ($request->btnSubmit === 'btnStore') {
+        if ($request->btnSubmit === 'btnSetPassword') {
+            $rc = $rc = redirect('/user-editpassword/' . strval($user->id));
+        } elseif ($request->btnSubmit === 'btnStore') {
             $fields = $request->all();
             $validator = Validator::make($fields, $this->rules(false, $user));
             if ($validator->fails()) {
